requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-spinner==0.1.10
colorama==0.4.6
contextlib2==21.6.0
croniter==1.4.1
dm-tree==0.1.8
exceptiongroup==1.2.0
execnet==2.0.2
fastapi==0.106.0
fastapi-utilities==0.2.0
Flask==3.0.0
flatbuffers==23.5.26
frozenlist==1.4.1
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
mock==5.1.0
multidict==6.0.4
namex==0.0.7
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
path==16.9.0
path.py==12.5.0
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pytest==7.4.3
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor==2.4.0
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
zipp==3.17.0
